pkg/types: drop go1.4 tests
